+++ src/handlers/records-query.ts
@@ -6,12 +6,13 @@ import type { Filter, GenericMessage, MessageSort } from '../types/message-types
 import type { RecordsQueryMessage, RecordsQueryReply, RecordsWriteMessageWithOptionalEncodedData } from '../types/records-types.js';
 
 import { authenticate } from '../core/auth.js';
+import { DateSort } from '../types/records-types.js';
 import { messageReplyFromError } from '../core/message-reply.js';
 import { ProtocolAuthorization } from '../core/protocol-authorization.js';
 import { Records } from '../utils/records.js';
 
+import { RecordsQuery } from '../interfaces/records-query.js';
 import { SortOrder } from '../types/message-types.js';
-import { DateSort, RecordsQuery } from '../interfaces/records-query.js';
 import { DwnInterfaceName, DwnMethodName } from '../enums/dwn-interface-method.js';
 
 export class RecordsQueryHandler implements MethodHandler {
